Feat: 删除前先检查通知器是否存在

+++ nuoche/src/main/java/me/ningyu/app/nuoche/controller/ProviderController.java
@@ -48,8 +48,11 @@ public class ProviderController
     @DeleteMapping("/{id}")
     public void delete(@PathVariable String id)
     {
-        providerService.get(id)
-        providerService.deleteById(id);
+        ProviderVO vo = providerService.get(id);
+        if (vo != null)
+        {
+            providerService.deleteById(id);
+        }
     }
 
     @PutMapping("/{id}")
